`useCurrentUrl` をContextを使う形に変更
